!!! FEATURE: Translatable forms
This change allows for forms to be translated using XLIFF as elsewhere
in Flow. Form element labels, placeholders, option labels and more can
be translated.

Instructions are included in a new documentation chapter.

This is marked as breaking: because if you changed the "translationPackage"
earlier to customize validation error messages, you will need to use the
new "validationErrorTranslationPackage" setting for that now. A code
migration to adjust that is included, run:

    ./flow flow:core:migrate  --version 20160601101500 --package-key ...